+++ src/models/Buyer.js
@@ -5,7 +5,8 @@ const buyerSchema = new mongoose.Schema({
   email: { type: String, required: true, unique: true },
   phoneNumber: { type: String, required: true },
   password: { type: String, required: true },
-  address: { type: String, required: true }
+  address: { type: String, required: true },
+  profilePicUrl:{ type: String,}
 });
 
 module.exports = mongoose.model("Buyer", buyerSchema);
