+++ src/routers/Auth/authRoutes.js
@@ -7,7 +7,7 @@ const {
   logoutUser 
 } = require("../controllers/authController");
 
-const authMiddleware = require("../middleware/authMiddleware");
+const authMiddleware = require("../../middleware/authMiddleware");
 
 const router = express.Router();
